% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{add_facets}
\alias{add_facets}
\title{Add faceting to a plot}
\usage{
add_facets(plot, facet_by, facet_scales, nrow, ncol, byrow, ...)
}
\arguments{
\item{plot}{ggplot object}

\item{facet_by}{Column(s) to facet by}

\item{facet_scales}{Scale type for facets}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{byrow}{Fill by row}

\item{...}{Additional arguments for facet functions}
}
\value{
Faceted plot
}
\description{
Add faceting to a plot
}
\keyword{internal}
