% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{adjcolors}
\alias{adjcolors}
\title{Adjust colors with alpha}
\usage{
adjcolors(colors, alpha)
}
\arguments{
\item{colors}{Color vector}

\item{alpha}{Alpha value}
}
\value{
Colors with alpha applied
}
\description{
Adjust colors with alpha
}
\keyword{internal}
