% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{blend_colors}
\alias{blend_colors}
\title{Blend multiple colors}
\usage{
blend_colors(colors, mode = c("blend", "average", "screen", "multiply"))
}
\arguments{
\item{colors}{Vector of colors}

\item{mode}{Blend mode}
}
\value{
Blended color
}
\description{
Blend multiple colors
}
\keyword{internal}
