% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-specs.R
\name{build_element_line}
\alias{build_element_line}
\title{Build element_line from style spec}
\usage{
build_element_line(spec_path, ...)
}
\arguments{
\item{spec_path}{Character. Path to style specification}

\item{...}{Additional arguments to override spec}
}
\value{
A ggplot2::element_line object
}
\description{
Converts a style specification to ggplot2::element_line()
}
\keyword{internal}
