% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-specs.R
\name{build_element_text}
\alias{build_element_text}
\title{Build element_text from style spec}
\usage{
build_element_text(spec_path, base_size = 12, ...)
}
\arguments{
\item{spec_path}{Character. Path to style specification}

\item{base_size}{Numeric. Base font size}

\item{...}{Additional arguments to override spec}
}
\value{
A ggplot2::element_text object
}
\description{
Converts a style specification to ggplot2::element_text()
}
\keyword{internal}
