% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plot-builder.R
\name{build_plot}
\alias{build_plot}
\title{Build a plot with standard workflow}
\usage{
build_plot(
  data,
  atomic_fn,
  params,
  split_by = NULL,
  facet_by = NULL,
  combine = TRUE,
  ...
)
}
\arguments{
\item{data}{Data frame}

\item{atomic_fn}{Function to create atomic plot}

\item{params}{List of parameters}

\item{split_by}{Column to split by}

\item{facet_by}{Column to facet by}

\item{combine}{Whether to combine plots}

\item{...}{Additional arguments}
}
\value{
Plot or list of plots
}
\description{
This is the core function that handles the standard plot workflow:
\enumerate{
\item Validate parameters
\item Split data if requested
\item Build atomic plots
\item Combine or facet plots
}
}
\keyword{internal}
