% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{calc_justification}
\alias{calc_justification}
\title{Calculate text justification based on angle}
\usage{
calc_justification(angle)
}
\arguments{
\item{angle}{Text angle in degrees}
}
\value{
List with h (hjust) and v (vjust) values
}
\description{
Calculate text justification based on angle
}
\keyword{internal}
