% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxviolinplot.R
\name{calculate_legend_dimensions}
\alias{calculate_legend_dimensions}
\title{Calculate legend dimensions}
\usage{
calculate_legend_dimensions(legend.position, legend.direction)
}
\arguments{
\item{legend.position}{Character string; legend position}

\item{legend.direction}{Character string; legend direction}
}
\value{
List with width and height additions
}
\description{
Calculates additional space needed for the legend based on its
position and direction.
}
\keyword{internal}
