% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plot-builder.R
\name{check_legend_param}
\alias{check_legend_param}
\title{Check and normalize legend parameter}
\usage{
check_legend_param(legend_param, split_names, param_name)
}
\arguments{
\item{legend_param}{Legend position or direction}

\item{split_names}{Names of splits}

\item{param_name}{Parameter name for error messages}
}
\value{
Normalized legend parameter
}
\description{
Check and normalize legend parameter
}
\keyword{internal}
