% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-palettes.R
\name{check_palcolor}
\alias{check_palcolor}
\title{Check and normalize palcolor argument}
\usage{
check_palcolor(palcolor, split_names)
}
\description{
Check and normalize palcolor argument
}
\keyword{internal}
