% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine multiple plots}
\usage{
combine_plots(
  plots,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = NULL,
  axes = NULL,
  axis_titles = NULL,
  guides = NULL,
  design = NULL
)
}
\arguments{
\item{plots}{List of plot objects}

\item{combine}{Whether to combine}

\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{byrow}{Fill by row}

\item{axes}{Axis handling}

\item{axis_titles}{Axis title handling}

\item{guides}{Guide handling}

\item{design}{Custom design}
}
\value{
Combined plot or list
}
\description{
Combine multiple plots
}
\keyword{internal}
