% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-config.R
\name{config}
\alias{config}
\title{ggforge Global Configuration}
\description{
Configuration system for ggforge package. This provides a centralized
way to manage default settings and options.
}
\keyword{internal}
