% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{create_bg_layer}
\alias{create_bg_layer}
\title{Create background layer for plots}
\usage{
create_bg_layer(
  data,
  x,
  palette,
  palcolor,
  alpha,
  keep_empty,
  facet_by,
  direction = "vertical"
)
}
\arguments{
\item{data}{Data frame}

\item{x}{X column}

\item{palette}{Palette name}

\item{palcolor}{Custom colors}

\item{alpha}{Alpha value}

\item{keep_empty}{Keep empty levels}

\item{facet_by}{Faceting columns}

\item{direction}{Background direction}
}
\value{
ggplot layer
}
\description{
Create background layer for plots
}
\keyword{internal}
