% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{detect_var_type}
\alias{detect_var_type}
\title{Detect variable type}
\usage{
detect_var_type(data, column)
}
\arguments{
\item{data}{Data frame}

\item{column}{Character. Column name to check}
}
\value{
Character. One of "continuous", "discrete", "temporal", "ordered"
}
\description{
Determines the type of a variable for styling purposes.

Type hierarchy:
\itemize{
\item continuous: numeric (not factor)
\item discrete: factor, character, logical
\item temporal: Date, POSIXct, POSIXlt
\item ordered: ordered factor
}
}
\examples{
\dontrun{
data <- data.frame(
  x = 1:10,
  y = factor(letters[1:10]),
  z = Sys.Date() + 1:10
)
detect_var_type(data, "x") # "continuous"
detect_var_type(data, "y") # "discrete"
detect_var_type(data, "z") # "temporal"
}
}
\keyword{internal}
