% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocityplot.R
\name{.build_grid_layer}
\alias{.build_grid_layer}
\title{Build grid velocity layer}
\usage{
.build_grid_layer(
  embedding,
  v_embedding,
  density,
  smooth,
  n_neighbors,
  min_mass,
  scale,
  arrow_angle,
  arrow_color,
  arrow_alpha
)
}
\description{
Build grid velocity layer
}
\keyword{internal}
