% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocityplot.R
\name{.build_raw_layer}
\alias{.build_raw_layer}
\title{Build raw velocity layer}
\usage{
.build_raw_layer(
  embedding,
  v_embedding,
  density,
  scale,
  group_by,
  group_name,
  group_palette,
  group_palcolor,
  arrow_angle,
  arrow_color,
  arrow_alpha
)
}
\description{
Build raw velocity layer
}
\keyword{internal}
