% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocityplot.R
\name{.build_stream_layer}
\alias{.build_stream_layer}
\title{Build stream velocity layer}
\usage{
.build_stream_layer(
  embedding,
  v_embedding,
  density,
  smooth,
  n_neighbors,
  min_mass,
  cutoff_perc,
  arrow_angle,
  arrow_color,
  streamline_l,
  streamline_minl,
  streamline_res,
  streamline_n,
  streamline_width,
  streamline_alpha,
  streamline_color,
  streamline_palette,
  streamline_palcolor,
  streamline_bg_color,
  streamline_bg_stroke
)
}
\description{
Build stream velocity layer
}
\keyword{internal}
