% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.cox_curve_plot}
\alias{.cox_curve_plot}
\title{Cox Curve Plot}
\usage{
.cox_curve_plot(
  data,
  var,
  time,
  event,
  ribbon_color,
  ribbon_alpha,
  line_color,
  line_type,
  line_width,
  text_size,
  text_digit,
  text_face,
  xlab,
  ylab,
  title,
  show_cindex,
  theme,
  theme_args,
  aspect.ratio
)
}
\description{
Cox Curve Plot
}
\keyword{internal}
