% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.cox_forest2_plot}
\alias{.cox_forest2_plot}
\title{Cox Forest Plot (Detailed)}
\usage{
.cox_forest2_plot(
  data,
  point_colors,
  point_size,
  point_border_width,
  point_border_size,
  line_colors,
  line_type,
  line_width,
  cutoff_vline_type,
  cutoff_vline_width,
  cutoff_vline_color,
  x_log10_scale,
  text_colors,
  digits,
  rel_width,
  theme,
  theme_args
)
}
\description{
Cox Forest Plot (Detailed)
}
\keyword{internal}
