% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.cox_forest_plot}
\alias{.cox_forest_plot}
\title{Cox Forest Plot (Simple)}
\usage{
.cox_forest_plot(
  data,
  point_colors,
  point_size,
  point_border_width,
  point_border_size,
  line_colors,
  line_type,
  line_width,
  cutoff_vline_type,
  cutoff_vline_width,
  cutoff_vline_color,
  x_log10_scale,
  xlab,
  title,
  subtitle,
  legend.position,
  legend.direction,
  theme,
  theme_args,
  aspect.ratio
)
}
\description{
Cox Forest Plot (Simple)
}
\keyword{internal}
