% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.create_forest_base}
\alias{.create_forest_base}
\title{Create base forest plot}
\usage{
.create_forest_base(
  data,
  point_colors,
  point_size,
  point_border_width,
  point_border_size,
  line_colors,
  line_type,
  line_width,
  cutoff_vline_type,
  cutoff_vline_width,
  cutoff_vline_color,
  x_log10_scale
)
}
\description{
Shared base plot creation for forest plots
}
\keyword{internal}
