% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.create_text_geom}
\alias{.create_text_geom}
\title{Create text geom for forest2}
\usage{
.create_text_geom(mapping, text_size, point_colors, text_colors, ...)
}
\description{
Helper to create text geoms with optional coloring
}
\keyword{internal}
