% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-palettes.R
\name{.get_continuous_colors}
\alias{.get_continuous_colors}
\title{Get continuous colors}
\usage{
.get_continuous_colors(x, palcolor, pal_n, n, matched, NA_color)
}
\description{
Get continuous colors
}
\keyword{internal}
