% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.get_cox_table}
\alias{.get_cox_table}
\title{Compute Cox Table}
\usage{
.get_cox_table(
  data,
  vars,
  time = "time",
  event = "event",
  scale = FALSE,
  nonExpression_ratio = 1,
  parallel = FALSE,
  n_cores = 1
)
}
\description{
Internal function to compute Cox proportional-hazards model summary table
with optional parallel computation
}
\keyword{internal}
