% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-palettes.R
\name{.get_discrete_colors}
\alias{.get_discrete_colors}
\title{Get discrete colors}
\usage{
.get_discrete_colors(x, palcolor, pal_n, matched, NA_color)
}
\description{
Get discrete colors
}
\keyword{internal}
