% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-config.R
\docType{data}
\name{.ggforge_defaults}
\alias{.ggforge_defaults}
\title{Default configuration values}
\format{
An object of class \code{list} of length 6.
}
\usage{
.ggforge_defaults
}
\description{
Default configuration values
}
\keyword{internal}
