% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-specs.R
\docType{data}
\name{.ggforge_style_specs}
\alias{.ggforge_style_specs}
\title{Core style specifications}
\format{
An object of class \code{list} of length 3.
}
\usage{
.ggforge_style_specs
}
\description{
Core style specifications
}
\keyword{internal}
