% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{.prepare_extent}
\alias{.prepare_extent}
\title{Prepare extent for spatial plots}
\usage{
.prepare_extent(ext)
}
\arguments{
\item{ext}{A numeric vector of length 4 or a SpatExtent object}
}
\value{
A SpatExtent object or NULL
}
\description{
Prepare extent for spatial plots
}
\keyword{internal}
