% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxplot.R
\name{.validate_cox_params}
\alias{.validate_cox_params}
\title{Validate Cox-specific parameters}
\usage{
.validate_cox_params(n_cores, digits, rel_width, nonExpression_ratio)
}
\description{
Validate Cox-specific parameters
}
\keyword{internal}
