% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialplot.R
\name{.wrap_spatial_layers}
\alias{.wrap_spatial_layers}
\title{Wrap spatial plot layers}
\usage{
.wrap_spatial_layers(
  layers,
  ext = NULL,
  flip_y = TRUE,
  legend.position = "right",
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  theme = "theme_ggforge",
  theme_args = list()
)
}
\arguments{
\item{layers}{List of ggplot layers}

\item{ext}{Spatial extent}

\item{flip_y}{Whether to flip y-axis}

\item{legend.position}{Legend position}

\item{legend.direction}{Legend direction}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{theme}{Theme function or name}

\item{theme_args}{Theme arguments}
}
\value{
A ggplot object
}
\description{
Wrap spatial plot layers
}
\keyword{internal}
