% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrich_example}
\alias{enrich_example}
\title{Example Enrichment Analysis Data}
\format{
A data frame with enrichment results
}
\usage{
enrich_example
}
\description{
Example data from clusterProfiler enrichment analysis.
Used for demonstrating enrichment plot functions.
}
\keyword{datasets}
