% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrich_multidb_example}
\alias{enrich_multidb_example}
\title{Example Multi-Database Enrichment Data}
\format{
A data frame with enrichment results from multiple databases
}
\usage{
enrich_multidb_example
}
\description{
Example enrichment data from multiple databases.
}
\keyword{datasets}
