% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plot-builder.R
\name{format_title}
\alias{format_title}
\title{Format title with dynamic content}
\usage{
format_title(title, split_name, default_title)
}
\arguments{
\item{title}{Title (string or function)}

\item{split_name}{Name of current split}

\item{default_title}{Default title}
}
\value{
Formatted title
}
\description{
Format title with dynamic content
}
\keyword{internal}
