% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-parameters.R
\name{get_default_params}
\alias{get_default_params}
\title{Get default plot parameters}
\usage{
get_default_params(plot_type = NULL)
}
\arguments{
\item{plot_type}{Optional plot type to get specific defaults}
}
\value{
List of default parameters
}
\description{
Returns a list of default parameters for all plot types
}
\keyword{internal}
