% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsea_example}
\alias{gsea_example}
\title{Example GSEA Results}
\format{
A data frame with GSEA results
}
\usage{
gsea_example
}
\description{
Example Gene Set Enrichment Analysis results.
}
\keyword{datasets}
