% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxviolinplot.R
\name{has_zero_variance}
\alias{has_zero_variance}
\title{Check if a vector has zero variance}
\usage{
has_zero_variance(vec)
}
\arguments{
\item{vec}{Numeric vector to check}
}
\value{
Logical; TRUE if vector has zero variance, FALSE otherwise
}
\description{
Determines whether a numeric vector has zero variance (all values identical
or all NA). Used to prevent statistical test failures.
}
\examples{
\dontrun{
has_zero_variance(c(1, 1, 1, 1))
# [1] TRUE

has_zero_variance(c(1, 2, 3, 4))
# [1] FALSE

has_zero_variance(c(NA, NA, NA))
# [1] TRUE

has_zero_variance(c(5, NA, 5, NA))
# [1] TRUE
}

}
\keyword{internal}
