% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{is_continuous_var}
\alias{is_continuous_var}
\title{Check if variable is continuous}
\usage{
is_continuous_var(data, column)
}
\arguments{
\item{data}{Data frame}

\item{column}{Character. Column name}
}
\value{
Logical
}
\description{
Helper to check if a variable should use continuous styling.
}
\keyword{internal}
