% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-themes.R
\name{map_type_to_style}
\alias{map_type_to_style}
\title{Map variable type to style specification path}
\usage{
map_type_to_style(type)
}
\arguments{
\item{type}{Character. Variable type from detect_var_type()}
}
\value{
Character. Style spec path
}
\description{
Map variable type to style specification path
}
\keyword{internal}
