% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-parameters.R
\name{merge_params}
\alias{merge_params}
\title{Merge plot parameters}
\usage{
merge_params(user_params, defaults)
}
\arguments{
\item{user_params}{User-provided parameters}

\item{defaults}{Default parameters}
}
\value{
Merged parameter list
}
\description{
Intelligently merge user-provided parameters with defaults
}
\keyword{internal}
