% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-parameters.R
\name{new_plot_params}
\alias{new_plot_params}
\title{Plot Parameter Class}
\usage{
new_plot_params(...)
}
\arguments{
\item{...}{Named parameters}
}
\value{
An object of class "plot_params"
}
\description{
S3 class for storing and validating plot parameters
}
\keyword{internal}
