% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{normalize_expansion}
\alias{normalize_expansion}
\title{Normalize expansion values (CSS-like padding)}
\usage{
normalize_expansion(
  expand,
  x_type,
  y_type,
  continuous_default = c(0.05, 0),
  discrete_default = c(0, 0.6)
)
}
\arguments{
\item{expand}{Numeric vector of expansion values}

\item{x_type}{Type of x-axis ("continuous" or "discrete")}

\item{y_type}{Type of y-axis ("continuous" or "discrete")}

\item{continuous_default}{Default for continuous axes}

\item{discrete_default}{Default for discrete axes}
}
\value{
List with x and y expansion values
}
\description{
Normalize expansion values (CSS-like padding)
}
\keyword{internal}
