% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{prepare_fgsea_result}
\alias{prepare_fgsea_result}
\title{Prepare fgsea result for plotting}
\usage{
prepare_fgsea_result(data)
}
\arguments{
\item{data}{A data frame of fgsea results}
}
\value{
A data frame with standardized column names for plotting
}
\description{
Converts fgsea package output format to DOSE-compatible format for plotting
}
\keyword{internal}
