% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-themes.R
\name{process_theme}
\alias{process_theme}
\title{Process theme argument}
\usage{
process_theme(theme)
}
\arguments{
\item{theme}{Theme name or function}
}
\value{
Theme function
}
\description{
Converts theme name string to theme function
}
\keyword{internal}
