% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{QQPlotAtomic}
\alias{QQPlotAtomic}
\title{QQ Plot Atomic}
\usage{
QQPlotAtomic(
  data,
  val,
  val_trans = NULL,
  type = "qq",
  band = NULL,
  line = list(),
  point = list(),
  fill_name = "Bands",
  band_alpha = 0.5,
  theme = "theme_ggforge",
  theme_args = list(),
  palette = "Spectral",
  palcolor = NULL,
  facet_by = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  facet_byrow = TRUE,
  aspect.ratio = 1,
  legend.position = waiver(),
  legend.direction = "vertical",
  title = NULL,
  subtitle = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  seed = 8525,
  ...
)
}
\arguments{
\item{data}{A data frame containing the data to plot}

\item{val}{A character string of the column name for the values to plot.
A numeric column is expected.}

\item{val_trans}{A function to transform the values before plotting.
Default is NULL, which means no transformation.}

\item{type}{A character string to specify the type of plot.
Default is "qq", which means QQ plot. Other option is "pp" for PP plot.}

\item{band}{A list of arguments to pass to \code{\link[qqplotr:stat_qq_band]{qqplotr::stat_qq_band()}} or
\code{\link[qqplotr:stat_pp_band]{qqplotr::stat_pp_band()}}, depending on the value of \code{type}.
Default is NULL, which means no band. If an empty list or TRUE is provided,
the default arguments will be used. Multiple bands can be added by providing
a list of lists.}

\item{line}{A list of arguments to pass to \code{\link[qqplotr:stat_qq_line]{qqplotr::stat_qq_line()}} or
\code{\link[qqplotr:stat_pp_line]{qqplotr::stat_pp_line()}}, depending on the value of \code{type}.
Default is \code{list()}, which means to add a line with default arguments.
If \code{NULL} is provided, no line will be added.}

\item{point}{A list of arguments to pass to \code{\link[qqplotr:stat_qq_point]{qqplotr::stat_qq_point()}} or
\code{\link[qqplotr:stat_pp_point]{qqplotr::stat_pp_point()}}, depending on the value of \code{type}.
Default is \code{list()}, which means to add points with default arguments.
If \code{NULL} is provided, no points will be added (not recommended).}

\item{fill_name}{A character string to name the legend of fill.
Default is "Bands".}

\item{band_alpha}{A numeric value to set the alpha of all bands.
Default is 0.5. It is a shortcut for setting alpha of all bands.
You can override it by setting \code{alpha} in \code{band} argument.}

\item{theme}{Theme name (string) or theme function}

\item{theme_args}{List of arguments passed to theme function}

\item{palette}{Color palette name}

\item{palcolor}{Custom colors for palette}

\item{facet_by}{Column name(s) for faceting the plot}

\item{facet_scales}{Scales for facets: "fixed", "free", "free_x", "free_y"}

\item{facet_nrow}{Number of rows in facet layout}

\item{facet_ncol}{Number of columns in facet layout}

\item{facet_byrow}{Fill facets by row (TRUE) or column (FALSE)}

\item{aspect.ratio}{Aspect ratio of plot panel}

\item{legend.position}{Legend position: "none", "left", "right", "bottom", "top"}

\item{legend.direction}{Legend direction: "horizontal" or "vertical"}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{xlim}{A numeric vector of length 2 to set the x-axis limits.}

\item{ylim}{A numeric vector of length 2 to set the y-axis limits.}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{seed}{Random seed for reproducibility}
}
\description{
Creates a single QQ or PP plot without splitting
}
\keyword{internal}
