% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeyplot.R
\name{SankeyPlot}
\alias{SankeyPlot}
\alias{AlluvialPlot}
\title{Sankey / Alluvial Plot}
\usage{
SankeyPlot(
  data,
  in_form = c("auto", "long", "lodes", "wide", "alluvia", "counts"),
  x,
  x_sep = "_",
  y = NULL,
  stratum = NULL,
  stratum_sep = "_",
  alluvium = NULL,
  alluvium_sep = "_",
  split_by = NULL,
  split_by_sep = "_",
  keep_empty = TRUE,
  flow = FALSE,
  expand = c(0, 0, 0, 0),
  nodes_legend = c("auto", "separate", "merge", "none"),
  nodes_color = "grey30",
  links_fill_by = NULL,
  links_fill_by_sep = "_",
  links_name = NULL,
  links_color = "gray80",
  nodes_palette = "Paired",
  nodes_palcolor = NULL,
  nodes_alpha = 1,
  nodes_label = FALSE,
  nodes_label_miny = 0,
  nodes_width = 0.25,
  links_palette = "Paired",
  links_palcolor = NULL,
  links_alpha = 0.6,
  legend.box = "vertical",
  x_text_angle = 0,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  flip = FALSE,
  theme = "theme_ggforge",
  theme_args = list(),
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_by = NULL,
  facet_scales = "fixed",
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_byrow = TRUE,
  seed = 8525,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  axes = NULL,
  axis_titles = axes,
  guides = NULL,
  design = NULL,
  ...
)

AlluvialPlot(
  data,
  in_form = c("auto", "long", "lodes", "wide", "alluvia", "counts"),
  x,
  x_sep = "_",
  y = NULL,
  stratum = NULL,
  stratum_sep = "_",
  alluvium = NULL,
  alluvium_sep = "_",
  split_by = NULL,
  split_by_sep = "_",
  keep_empty = TRUE,
  flow = FALSE,
  expand = c(0, 0, 0, 0),
  nodes_legend = c("auto", "separate", "merge", "none"),
  nodes_color = "grey30",
  links_fill_by = NULL,
  links_fill_by_sep = "_",
  links_name = NULL,
  links_color = "gray80",
  nodes_palette = "Paired",
  nodes_palcolor = NULL,
  nodes_alpha = 1,
  nodes_label = FALSE,
  nodes_label_miny = 0,
  nodes_width = 0.25,
  links_palette = "Paired",
  links_palcolor = NULL,
  links_alpha = 0.6,
  legend.box = "vertical",
  x_text_angle = 0,
  aspect.ratio = 1,
  legend.position = "right",
  legend.direction = "vertical",
  flip = FALSE,
  theme = "theme_ggforge",
  theme_args = list(),
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_by = NULL,
  facet_scales = "fixed",
  facet_ncol = NULL,
  facet_nrow = NULL,
  facet_byrow = TRUE,
  seed = 8525,
  combine = TRUE,
  nrow = NULL,
  ncol = NULL,
  byrow = TRUE,
  axes = NULL,
  axis_titles = axes,
  guides = NULL,
  design = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame in following possible formats:
\itemize{
\item "long" or "lodes": A long format with columns for \code{x}, \code{stratum}, \code{alluvium}, and \code{y}.
\code{x} (required, single columns or concatenated by \code{x_sep}) is the column name to plot on the x-axis,
\code{stratum} (defaults to \code{links_fill_by}) is the column name
to group the nodes for each \code{x}, \code{alluvium} (required) is the column name to define the links, and \code{y}
is the frequency of each \code{x}, \code{stratum}, and \code{alluvium}.
\item "wide" or "alluvia": A wide format with columns for \code{x}.
\code{x} (required, multiple columns, \code{x_sep} won't be used) are the columns to plot on the x-axis,
\code{stratum} and \code{alluvium} will be ignored.
See \link[ggalluvial:alluvial-data]{ggalluvial::to_lodes_form} for more details.
\item "counts": A format with counts being provides under each \code{x}.
\code{x} (required, multiple columns, \code{x_sep} won't be used) are the columns to plot on the x-axis.
When the first element of \code{x} is ".", values of \code{links_fill_by} (required) will be added to the plot as the first column of nodes.
It is useful to show how the links are flowed from the source to the targets.
\item "auto" (default): Automatically determine the format based on the columns provided.
When the length of \code{x} is greater than 1 and all \code{x} columns are numeric, "counts" format will be used.
When the length of \code{x} is greater than 1 and \link[ggalluvial:alluvial-data]{ggalluvial::is_alluvia_form} returns TRUE, "alluvia" format will be used.
Otherwise, "lodes" format will be tried.
}}

\item{in_form}{A character string to specify the format of the data.
Possible values are "auto", "long", "lodes", "wide", "alluvia", and "counts".}

\item{x}{A character string of the column name to plot on the x-axis.
See \code{data} for more details.}

\item{x_sep}{A character string to concatenate the columns in \code{x}, if multiple columns are provided.}

\item{y}{A character string of the column name to plot on the y-axis.
When \code{in_form} is "counts", \code{y} will be ignored. Otherwise, it defaults to the count of each \code{x}, \code{stratum}, \code{alluvium} and \code{links_fill_by}.}

\item{stratum}{A character string of the column name to group the nodes for each \code{x}.
See \code{data} for more details.}

\item{stratum_sep}{A character string to concatenate the columns in \code{stratum}, if multiple columns are provided.}

\item{alluvium}{A character string of the column name to define the links.
See \code{data} for more details.}

\item{alluvium_sep}{A character string to concatenate the columns in \code{alluvium}, if multiple columns are provided.}

\item{split_by}{Column name(s) to split data into multiple plots}

\item{split_by_sep}{Separator when concatenating multiple split_by columns}

\item{keep_empty}{A logical value to keep the empty nodes.}

\item{flow}{A logical value to use \link[ggalluvial:geom_flow]{ggalluvial::geom_flow} instead of \link[ggalluvial:geom_alluvium]{ggalluvial::geom_alluvium}.}

\item{expand}{Expansion values for axes (CSS-like: top, right, bottom, left)}

\item{nodes_legend}{Controls how the legend of nodes will be shown. Possible values are:
\itemize{
\item "merge": Merge the legends of nodes. That is only one legend will be shown for all nodes.
\item "separate": Show the legends of nodes separately. That is, nodes on each \code{x} will have their own legend.
\item "none": Do not show the legend of nodes.
\item "auto": Automatically determine how to show the legend.
When \code{nodes_label} is TRUE, "none" will apply.
When \code{nodes_label} is FALSE, and if stratum is the same as links_fill_by, "none" will apply.
If there is any overlapping values between the nodes on different \code{x},
"merge" will apply. Otherwise, "separate" will apply.
}}

\item{nodes_color}{A character string to color the nodes.
Use a special value ".fill" to use the same color as the fill.}

\item{links_fill_by}{A character string of the column name to fill the links.}

\item{links_fill_by_sep}{A character string to concatenate the columns in \code{links_fill_by}, if multiple columns are provided.}

\item{links_name}{A character string to name the legend of links.}

\item{links_color}{A character string to color the borders of links.
Use a special value ".fill" to use the same color as the fill.}

\item{nodes_palette}{A character string to specify the palette of nodes fill.}

\item{nodes_palcolor}{A character vector to specify the colors of nodes fill.}

\item{nodes_alpha}{A numeric value to specify the transparency of nodes fill.}

\item{nodes_label}{A logical value to show the labels on the nodes.}

\item{nodes_label_miny}{A numeric value to specify the minimum y (frequency) to show the labels.}

\item{nodes_width}{A numeric value to specify the width of nodes.}

\item{links_palette}{A character string to specify the palette of links fill.}

\item{links_palcolor}{A character vector to specify the colors of links fill.}

\item{links_alpha}{A numeric value to specify the transparency of links fill.}

\item{legend.box}{A character string to specify the box of the legend, either "vertical" or "horizontal".}

\item{x_text_angle}{Angle for x-axis text}

\item{aspect.ratio}{Aspect ratio of plot panel}

\item{legend.position}{Legend position: "none", "left", "right", "bottom", "top"}

\item{legend.direction}{Legend direction: "horizontal" or "vertical"}

\item{flip}{A logical value to flip the plot.}

\item{theme}{Theme name (string) or theme function}

\item{theme_args}{List of arguments passed to theme function}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{facet_by}{Column name(s) for faceting the plot}

\item{facet_scales}{Scales for facets: "fixed", "free", "free_x", "free_y"}

\item{facet_ncol}{Number of columns in facet layout}

\item{facet_nrow}{Number of rows in facet layout}

\item{facet_byrow}{Fill facets by row (TRUE) or column (FALSE)}

\item{seed}{Random seed for reproducibility}

\item{combine}{Whether to combine split plots into one}

\item{nrow}{Number of rows when combining plots}

\item{ncol}{Number of columns when combining plots}

\item{byrow}{Fill combined plots by row}

\item{axes}{How to handle axes in combined plots ("keep", "collect", "collect_x", "collect_y")}

\item{axis_titles}{How to handle axis titles in combined plots}

\item{guides}{How to handle guides in combined plots ("collect", "keep", "auto")}

\item{design}{Custom layout design for combined plots}

\item{...}{Other arguments to pass to \link[ggalluvial:geom_alluvium]{ggalluvial::geom_alluvium} or \link[ggalluvial:geom_flow]{ggalluvial::geom_flow}.}
}
\value{
A ggplot object or wrap_plots object or a list of ggplot objects
}
\description{
A plot visualizing flow/movement/change from one state to another or one time to another.
\code{AlluvialPlot} is an alias of \code{SankeyPlot}.
}
\examples{
\donttest{
# Reproduce the examples in ggalluvial
set.seed(8525)

data(UCBAdmissions, package = "datasets")
UCBAdmissions <- as.data.frame(UCBAdmissions)
SankeyPlot(as.data.frame(UCBAdmissions),
  x = c("Gender", "Dept"),
  y = "Freq", nodes_width = 1 / 12, links_fill_by = "Admit", nodes_label = TRUE,
  nodes_palette = "simspec", links_palette = "Set1", links_alpha = 0.5,
  nodes_palcolor = "black", links_color = "transparent"
)

data(HairEyeColor, package = "datasets")
SankeyPlot(as.data.frame(HairEyeColor),
  x = c("Hair", "Eye", "Sex"),
  y = "Freq", links_fill_by = "Eye", nodes_width = 1 / 8, nodes_alpha = 0.4,
  flip = TRUE, reverse = FALSE, knot.pos = 0, links_color = "transparent",
  ylab = "Freq", links_alpha = 0.5, links_name = "Eye (links)", links_palcolor = c(
    Brown = "#70493D", Hazel = "#E2AC76", Green = "#3F752B", Blue = "#81B0E4"
  )
)

data(Refugees, package = "alluvial")
country_regions <- c(
  Afghanistan = "Middle East",
  Burundi = "Central Africa",
  `Congo DRC` = "Central Africa",
  Iraq = "Middle East",
  Myanmar = "Southeast Asia",
  Palestine = "Middle East",
  Somalia = "Horn of Africa",
  Sudan = "Central Africa",
  Syria = "Middle East",
  Vietnam = "Southeast Asia"
)
Refugees$region <- country_regions[Refugees$country]
SankeyPlot(Refugees,
  x = "year", y = "refugees", alluvium = "country",
  links_fill_by = "country", links_color = ".fill", links_alpha = 0.75,
  links_palette = "Set3", facet_by = "region", x_text_angle = -45, nodes_legend = "none",
  theme_args = list(strip.background = ggplot2::element_rect(fill = "grey80")),
  decreasing = FALSE, nodes_width = 0, nodes_color = "transparent", ylab = "refugees",
  title = "Refugee volume by country and region of origin"
)

data(majors, package = "ggalluvial")
majors$curriculum <- as.factor(majors$curriculum)
SankeyPlot(majors,
  x = "semester", stratum = "curriculum", alluvium = "student",
  links_fill_by = "curriculum", flow = TRUE, stat = "alluvium", nodes_palette = "Set2",
  links_palette = "Set2"
)

data(vaccinations, package = "ggalluvial")
vaccinations <- transform(vaccinations,
  response = factor(response, rev(levels(response)))
)
SankeyPlot(vaccinations,
  x = "survey", stratum = "response", alluvium = "subject",
  y = "freq", links_fill_by = "response", nodes_label = TRUE, nodes_alpha = 0.5,
  nodes_palette = "seurat", links_palette = "seurat", links_alpha = 0.5,
  legend.position = "none", flow = TRUE, expand = c(0, 0, 0, .15), stat = "alluvium",
  title = "vaccination survey responses at three points in time"
)

data(Titanic, package = "datasets")
SankeyPlot(as.data.frame(Titanic),
  x = c("Class", "Sex"), y = "Freq",
  links_fill_by = "Survived", flow = TRUE, facet_by = "Age", facet_scales = "free_y",
  nodes_label = TRUE, expand = c(0.05, 0), xlab = "", links_palette = "Set1",
  nodes_palcolor = "white", nodes_label_miny = 10
)

# Simulated examples
df <- data.frame(
  Clone = paste0("clone", 1:10),
  Timepoint1 = sample(c(rep(0, 30), 1:100), 10),
  Timepoint2 = sample(c(rep(0, 30), 1:100), 10)
)
SankeyPlot(df,
  x = c("Timepoint1", "Timepoint2"), alluvium = "Clone",
  links_color = ".fill"
)

df <- data.frame(
  Clone = rep(paste0("clone", 1:6), each = 2),
  Timepoint1 = sample(c(rep(0, 30), 1:100), 6),
  Timepoint2 = sample(c(rep(0, 30), 1:100), 6),
  Group = rep(c("A", "B"), 6)
)
SankeyPlot(df,
  x = c(".", "Timepoint1", "Timepoint2"),
  stratum = "Group", links_fill_by = "Clone", links_color = ".fill"
)
}
}
