% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-utils.R
\name{split_data}
\alias{split_data}
\title{Split data by column}
\usage{
split_data(data, split_by)
}
\arguments{
\item{data}{Data frame}

\item{split_by}{Column to split by}
}
\value{
Named list of data frames
}
\description{
Split data by column
}
\keyword{internal}
