% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-themes.R
\name{theme_minimal_axes}
\alias{theme_minimal_axes}
\title{Minimal Theme}
\usage{
theme_minimal_axes(
  add_coord = TRUE,
  xlen_npc = 0.15,
  ylen_npc = 0.15,
  xlab = "",
  ylab = "",
  lab_size = 12,
  ...
)
}
\arguments{
\item{add_coord}{Whether to add coordinate arrows}

\item{xlen_npc}{Length of x-axis arrow (in npc units)}

\item{ylen_npc}{Length of y-axis arrow (in npc units)}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{lab_size}{Label size}

\item{...}{Additional arguments passed to theme}
}
\value{
List of ggplot2 theme components
}
\description{
A minimal theme with coordinate axes
}
