% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-validator.R
\name{validate_choice}
\alias{validate_choice}
\title{Validate choice from options}
\usage{
validate_choice(value, param_name, choices, allow_null = TRUE)
}
\arguments{
\item{value}{Value to validate}

\item{param_name}{Parameter name}

\item{choices}{Valid choices}

\item{allow_null}{Whether NULL is allowed}
}
\value{
The validated value
}
\description{
Validate choice from options
}
\keyword{internal}
