% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-types.R
\name{var_type}
\alias{var_type}
\title{Annotate variable with explicit type}
\usage{
var_type(x, type = c("continuous", "discrete", "temporal", "ordered"))
}
\arguments{
\item{x}{Vector. The variable to annotate}

\item{type}{Character. One of "continuous", "discrete", "temporal", "ordered"}
}
\value{
The vector with type attribute attached
}
\description{
Allows users to override automatic type detection.
}
\examples{
\dontrun{
# Force a numeric year column to be treated as discrete
data$year <- var_type(data$year, "discrete")

# Force a character column to be treated as continuous (rare)
data$score <- var_type(as.numeric(data$score), "continuous")
}
}
