% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-plot-builder.R
\name{wrap_atomic_plot}
\alias{wrap_atomic_plot}
\title{Standard atomic plot wrapper}
\usage{
wrap_atomic_plot(data, plot_fn, params)
}
\arguments{
\item{data}{Data frame}

\item{plot_fn}{Function to create the plot}

\item{params}{Parameters}
}
\value{
ggplot object
}
\description{
Wraps atomic plot functions with consistent pre/post processing
}
\keyword{internal}
